Fix #57 Skip already curated names
We assume here that name that was once curated will always
have the same semantic meaning in the future. While it is not
alloway so it is a goor rule of thumb that significantly
cuts time of a text curation. In the future we are going
to make it a command line option instead of runtime
option, because it is too confusing as a runtime.

Works for gnfinder's infraspecies detection, typified labels
(instead of strings), gnfinder training with using texts.